+++ src/Algorithms/ProximalGradientDescent.cpp
@@ -138,12 +138,12 @@ void ProximalGradientDescent::run(Gflasso * model) {
     double residue = model->cost();
     double theta = 1;
     double theta_new = 0;
-    MatrixXd beta_prev = model->get_beta(); //bx
-    MatrixXd beta_curr = model->get_beta(); //bx_new
-    MatrixXd beta = model->get_beta();  //bw
-    MatrixXd best_beta = model->get_beta();
-    MatrixXd in;
-    MatrixXd grad;
+    MatrixXf beta_prev = model->get_beta(); //bx
+    MatrixXf beta_curr = model->get_beta(); //bx_new
+    MatrixXf beta = model->get_beta();  //bw
+    MatrixXf best_beta = model->get_beta();
+    MatrixXf in;
+    MatrixXf grad;
     double diff = tolerance*2;
     prev_residue= 9999999;
     while (!shouldStop && epoch < maxIteration && diff > tolerance) {
@@ -169,7 +169,7 @@ void ProximalGradientDescent::run(Gflasso * model) {
 
 void ProximalGradientDescent::run(LinearRegression *model) {
     int epoch = 0;
-    MatrixXd y = model->getY();
+    MatrixXf y = model->getY();
     model->setL1_reg(model->getL1_reg()*10);
     long s = y.cols();
     for (long i=0; i<s; i++){
@@ -204,12 +204,12 @@ void ProximalGradientDescent::run(TreeLasso * model) {
     double residue = model->cost();
     double theta = 1;
     double theta_new = 0;
-    MatrixXd beta_prev = model->getBeta(); //bx
-    MatrixXd beta_curr = model->getBeta(); //bx_new
-    MatrixXd beta = model->getBeta();  //bw
-    MatrixXd best_beta = model->getBeta();
-    MatrixXd in;
-    MatrixXd grad;
+    MatrixXf beta_prev = model->getBeta(); //bx
+    MatrixXf beta_curr = model->getBeta(); //bx_new
+    MatrixXf beta = model->getBeta();  //bw
+    MatrixXf best_beta = model->getBeta();
+    MatrixXf in;
+    MatrixXf grad;
     model->initGradientUpdate();
     double diff = tolerance*2;
     while (!shouldStop && epoch < maxIteration && diff > tolerance) {
@@ -238,8 +238,8 @@ void ProximalGradientDescent::run(TreeLasso * model) {
 
 
 void ProximalGradientDescent::run(MultiPopLasso * model) {
-    MatrixXd X = model->getX();
-    MatrixXd y = model->getY();
+    MatrixXf X = model->getX();
+    MatrixXf y = model->getY();
     int epoch = 0;
     long s = y.cols();
     for (long i=0; i<y.cols(); i++) {
@@ -253,12 +253,12 @@ void ProximalGradientDescent::run(MultiPopLasso * model) {
         double residue = model->cost();
         double theta = 1;
         double theta_new = 0;
-        MatrixXd beta_prev = model->getFormattedBeta(); //bx
-        MatrixXd beta_curr = model->getFormattedBeta(); //bx_new
-        MatrixXd beta = model->getFormattedBeta();  //bw
-        MatrixXd best_beta = model->getFormattedBeta();
-        MatrixXd in;
-        MatrixXd grad;
+        MatrixXf beta_prev = model->getFormattedBeta(); //bx
+        MatrixXf beta_curr = model->getFormattedBeta(); //bx_new
+        MatrixXf beta = model->getFormattedBeta();  //bw
+        MatrixXf best_beta = model->getFormattedBeta();
+        MatrixXf in;
+        MatrixXf grad;
         double diff = tolerance * 2;
         while (!shouldStop && epoch < maxIteration && diff > tolerance) {
             epoch++;
@@ -292,13 +292,13 @@ void ProximalGradientDescent::run(AdaMultiLasso *model) {
     double residue = model->cost();
     double theta = 1;
     double theta_new = 0;
-    MatrixXd beta_prev = model->getFormattedBeta(); //bx
-    MatrixXd beta_curr = model->getFormattedBeta(); //bx_new
-    MatrixXd beta = model->getFormattedBeta();  //bw
-    MatrixXd best_beta = model->getFormattedBeta();
-    MatrixXd beta_prev2 = model->getFormattedBeta();
-    MatrixXd in;
-    MatrixXd grad;
+    MatrixXf beta_prev = model->getFormattedBeta(); //bx
+    MatrixXf beta_curr = model->getFormattedBeta(); //bx_new
+    MatrixXf beta = model->getFormattedBeta();  //bw
+    MatrixXf best_beta = model->getFormattedBeta();
+    MatrixXf beta_prev2 = model->getFormattedBeta();
+    MatrixXf in;
+    MatrixXf grad;
     double diff = tolerance*2;
     double lr2 = 0;
     long i1 = 0;
@@ -371,6 +371,6 @@ void ProximalGradientDescent::run(SparseLMM *model) {
     lr.setX(model->getRotatedX());
     lr.setY(model->getRoattedY());
     run(&lr);
-    MatrixXd tmp = lr.getBeta();
+    MatrixXf tmp = lr.getBeta();
     model->updateBeta(tmp);
 }
